Dates must not have a timezone specified but use the default timezone instead
Most of the functions are based on the knowledge of what is "today",
which is obtained from the browser using `new Date()`. This imply that
the dates used have been generated/manipulated from the browser, and
don't come from elsewhere with a different timezone. This IS a big
limitation, and I still didn't find a reliable solution.